+++ trs/admin.py
@@ -16,7 +16,8 @@ class PersonChangeAdmin(admin.ModelAdmin):
 
 
 class BookingAdmin(admin.ModelAdmin):
-    pass
+    list_display = ['year_week', 'hours', 'booked_by', 'booked_on']
+    list_filter = ['booked_by', 'booked_on']
 
 
 class WorkAssignmentAdmin(admin.ModelAdmin):
@@ -27,9 +28,15 @@ class BudgetAssignmentAdmin(admin.ModelAdmin):
     pass
 
 
+class YearWeekAdmin(admin.ModelAdmin):
+    # This one can be removed later on.
+    pass
+
+
 admin.site.register(models.Person, PersonAdmin)
 admin.site.register(models.Project, ProjectAdmin)
 admin.site.register(models.PersonChange, PersonChangeAdmin)
 admin.site.register(models.Booking, BookingAdmin)
 admin.site.register(models.WorkAssignment, WorkAssignmentAdmin)
 admin.site.register(models.BudgetAssignment, BudgetAssignmentAdmin)
+admin.site.register(models.YearWeek, YearWeekAdmin)