+++ trs/migrations/0014_auto__add_index_yearweek_first_day.py
@@ -1,0 +1,123 @@
+# -*- coding: utf-8 -*-
+import datetime
+from south.db import db
+from south.v2 import SchemaMigration
+from django.db import models
+
+
+class Migration(SchemaMigration):
+
+    def forwards(self, orm):
+        # Adding index on 'YearWeek', fields ['first_day']
+        db.create_index('trs_yearweek', ['first_day'])
+
+
+    def backwards(self, orm):
+        # Removing index on 'YearWeek', fields ['first_day']
+        db.delete_index('trs_yearweek', ['first_day'])
+
+
+    models = {
+        'auth.group': {
+            'Meta': {'object_name': 'Group'},
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'name': ('django.db.models.fields.CharField', [], {'max_length': '80', 'unique': 'True'}),
+            'permissions': ('django.db.models.fields.related.ManyToManyField', [], {'symmetrical': 'False', 'to': "orm['auth.Permission']", 'blank': 'True'})
+        },
+        'auth.permission': {
+            'Meta': {'unique_together': "(('content_type', 'codename'),)", 'ordering': "('content_type__app_label', 'content_type__model', 'codename')", 'object_name': 'Permission'},
+            'codename': ('django.db.models.fields.CharField', [], {'max_length': '100'}),
+            'content_type': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['contenttypes.ContentType']"}),
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'name': ('django.db.models.fields.CharField', [], {'max_length': '50'})
+        },
+        'auth.user': {
+            'Meta': {'object_name': 'User'},
+            'date_joined': ('django.db.models.fields.DateTimeField', [], {'default': 'datetime.datetime.now'}),
+            'email': ('django.db.models.fields.EmailField', [], {'max_length': '75', 'blank': 'True'}),
+            'first_name': ('django.db.models.fields.CharField', [], {'max_length': '30', 'blank': 'True'}),
+            'groups': ('django.db.models.fields.related.ManyToManyField', [], {'symmetrical': 'False', 'to': "orm['auth.Group']", 'blank': 'True', 'related_name': "'user_set'"}),
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'is_active': ('django.db.models.fields.BooleanField', [], {'default': 'True'}),
+            'is_staff': ('django.db.models.fields.BooleanField', [], {'default': 'False'}),
+            'is_superuser': ('django.db.models.fields.BooleanField', [], {'default': 'False'}),
+            'last_login': ('django.db.models.fields.DateTimeField', [], {'default': 'datetime.datetime.now'}),
+            'last_name': ('django.db.models.fields.CharField', [], {'max_length': '30', 'blank': 'True'}),
+            'password': ('django.db.models.fields.CharField', [], {'max_length': '128'}),
+            'user_permissions': ('django.db.models.fields.related.ManyToManyField', [], {'symmetrical': 'False', 'to': "orm['auth.Permission']", 'blank': 'True', 'related_name': "'user_set'"}),
+            'username': ('django.db.models.fields.CharField', [], {'max_length': '30', 'unique': 'True'})
+        },
+        'contenttypes.contenttype': {
+            'Meta': {'unique_together': "(('app_label', 'model'),)", 'ordering': "('name',)", 'db_table': "'django_content_type'", 'object_name': 'ContentType'},
+            'app_label': ('django.db.models.fields.CharField', [], {'max_length': '100'}),
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'model': ('django.db.models.fields.CharField', [], {'max_length': '100'}),
+            'name': ('django.db.models.fields.CharField', [], {'max_length': '100'})
+        },
+        'trs.booking': {
+            'Meta': {'object_name': 'Booking'},
+            'added': ('django.db.models.fields.DateTimeField', [], {'blank': 'True', 'auto_now_add': 'True'}),
+            'added_by': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['auth.User']", 'null': 'True', 'blank': 'True'}),
+            'booked_by': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['trs.Person']", 'null': 'True', 'blank': 'True', 'related_name': "'bookings'"}),
+            'booked_on': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['trs.Project']", 'null': 'True', 'blank': 'True', 'related_name': "'bookings'"}),
+            'hours': ('django.db.models.fields.DecimalField', [], {'decimal_places': '2', 'max_digits': '8', 'null': 'True', 'blank': 'True'}),
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'year_week': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['trs.YearWeek']", 'null': 'True', 'blank': 'True'})
+        },
+        'trs.budgetassignment': {
+            'Meta': {'object_name': 'BudgetAssignment'},
+            'added': ('django.db.models.fields.DateTimeField', [], {'blank': 'True', 'auto_now_add': 'True'}),
+            'added_by': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['auth.User']", 'null': 'True', 'blank': 'True'}),
+            'assigned_to': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['trs.Project']", 'null': 'True', 'blank': 'True', 'related_name': "'budget_assignments'"}),
+            'budget': ('django.db.models.fields.DecimalField', [], {'decimal_places': '2', 'max_digits': '12', 'null': 'True', 'blank': 'True'}),
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'year_week': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['trs.YearWeek']", 'null': 'True', 'blank': 'True'})
+        },
+        'trs.person': {
+            'Meta': {'ordering': "['name']", 'object_name': 'Person'},
+            'description': ('django.db.models.fields.CharField', [], {'max_length': '255', 'blank': 'True'}),
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'login_name': ('django.db.models.fields.CharField', [], {'max_length': '255'}),
+            'name': ('django.db.models.fields.CharField', [], {'max_length': '255'}),
+            'slug': ('django.db.models.fields.SlugField', [], {'max_length': '50'}),
+            'user': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['auth.User']", 'null': 'True', 'blank': 'True'})
+        },
+        'trs.personchange': {
+            'Meta': {'object_name': 'PersonChange'},
+            'added': ('django.db.models.fields.DateTimeField', [], {'blank': 'True', 'auto_now_add': 'True'}),
+            'added_by': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['auth.User']", 'null': 'True', 'blank': 'True'}),
+            'hours_per_week': ('django.db.models.fields.DecimalField', [], {'decimal_places': '2', 'max_digits': '8', 'null': 'True', 'blank': 'True'}),
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'person': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['trs.Person']", 'null': 'True', 'blank': 'True', 'related_name': "'person_changes'"}),
+            'target': ('django.db.models.fields.DecimalField', [], {'decimal_places': '2', 'max_digits': '8', 'null': 'True', 'blank': 'True'}),
+            'year_week': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['trs.YearWeek']", 'null': 'True', 'blank': 'True'})
+        },
+        'trs.project': {
+            'Meta': {'ordering': "['code']", 'object_name': 'Project'},
+            'added': ('django.db.models.fields.DateTimeField', [], {'blank': 'True', 'auto_now_add': 'True'}),
+            'code': ('django.db.models.fields.CharField', [], {'max_length': '255'}),
+            'description': ('django.db.models.fields.CharField', [], {'max_length': '255', 'blank': 'True'}),
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'slug': ('django.db.models.fields.SlugField', [], {'max_length': '50'})
+        },
+        'trs.workassignment': {
+            'Meta': {'object_name': 'WorkAssignment'},
+            'added': ('django.db.models.fields.DateTimeField', [], {'blank': 'True', 'auto_now_add': 'True'}),
+            'added_by': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['auth.User']", 'null': 'True', 'blank': 'True'}),
+            'assigned_on': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['trs.Project']", 'null': 'True', 'blank': 'True', 'related_name': "'work_assignments'"}),
+            'assigned_to': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['trs.Person']", 'null': 'True', 'blank': 'True', 'related_name': "'work_assignments'"}),
+            'hourly_tariff': ('django.db.models.fields.DecimalField', [], {'decimal_places': '2', 'max_digits': '8', 'null': 'True', 'blank': 'True'}),
+            'hours': ('django.db.models.fields.DecimalField', [], {'decimal_places': '2', 'max_digits': '8', 'null': 'True', 'blank': 'True'}),
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'year_week': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['trs.YearWeek']", 'null': 'True', 'blank': 'True'})
+        },
+        'trs.yearweek': {
+            'Meta': {'ordering': "['year', 'week']", 'object_name': 'YearWeek'},
+            'first_day': ('django.db.models.fields.DateField', [], {'db_index': 'True'}),
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'week': ('django.db.models.fields.IntegerField', [], {}),
+            'year': ('django.db.models.fields.IntegerField', [], {})
+        }
+    }
+
+    complete_apps = ['trs']