+++ CHANGES.rst
@@ -5,4 +5,9 @@ Changelog of trs
 0.1 (unreleased)
 ----------------
 
+- Added booking page including actual booking.
+
+- Added the initial set of models and base overview pages for
+  persons/projects.
+
 - Initial project structure created with nensskel 1.34.dev0.